Fix source map support
Since we use prepareStackTrace to capture the raw data for finding
call sites, we aren't actually getting sourcemapped information.
When the `source-map-support` package is available, it will be
loaded and used to process stack frames. We don't depend on it
directly, so we can't provide any compatibility guarantees; we just
fall back to the same behavior as before and print a warning if
this actually breaks.